Move image.operations -> image.volume_math and utils.image -> utils.volume_operations
