+++ scripts/scil_connectivity_math.py
@@ -17,7 +17,7 @@ import os
 import nibabel as nib
 import numpy as np
 
-from scilpy.image.operations import (get_array_ops, get_operations_doc)
+from scilpy.image.volume_math import (get_array_ops, get_operations_doc)
 from scilpy.io.utils import (add_overwrite_arg,
                              add_verbose_arg,
                              assert_outputs_exist,
