+++ scripts/scil_reshape_to_reference.py
@@ -17,7 +17,7 @@ import numpy as np
 
 from scilpy.io.utils import (add_overwrite_arg, assert_inputs_exist,
                              assert_outputs_exist)
-from scilpy.utils.image import transform_anatomy
+from scilpy.image.volume_operations import apply_transform
 
 
 def _build_arg_parser():
@@ -50,9 +50,9 @@ def main():
     assert_inputs_exist(parser, [args.in_file, args.in_ref_file])
     assert_outputs_exist(parser, args, args.out_file)
 
-    transform_anatomy(np.eye(4), args.in_ref_file, args.in_file,
-                      args.out_file, interp=args.interpolation,
-                      keep_dtype=args.keep_dtype)
+    apply_transform(np.eye(4), args.in_ref_file, args.in_file,
+                    args.out_file, interp=args.interpolation,
+                    keep_dtype=args.keep_dtype)
 
 
 if __name__ == "__main__":
